#42 - Add Deleted Test Fixture

+++ generators/app-symlink/templates/scripts/npm/create-app-symlink.sh
@@ -4,5 +4,8 @@ if [ ! -e node_modules ]; then
     mkdir -p node_modules
 fi
 
-rm node_modules/app
+if [ -e node_modules/app ]; then
+  rm node_modules/app
+fi
+
 ln -sf `pwd`/server node_modules/app
