+++ generators/skeleton/templates/scripts/npm/create-public-symlink.sh
@@ -4,5 +4,8 @@ if [ ! -e node_modules ]; then
     mkdir -p node_modules
 fi
 
-rm node_modules/public
+if [ -e node_modules/public ]; then
+  rm node_modules/public
+fi
+
 ln -sf `pwd`/public node_modules
